requirements.txt
aim==3.17.5
comet-ml==3.33.6
configobj==5.0.8
flake8==6.0.0
fqdn==1.5.1
hydra-core==1.3.2
librosa==0.10.0.post2
pip==23.1.2
pipdeptree==2.13.0
plotly==5.15.0
python-dotenv==1.0.0
pytorch-lightning==2.0.6
tensorboard==2.13.0
torchaudio==2.0.2
torchvision==0.15.2
webcolors==1.13